[Fix]フィールド追加処理修正
