+++ app/javascript/packs/application.js
@@ -2,7 +2,6 @@ import Rails from "@rails/ujs"
 import Turbolinks from "turbolinks"
 import * as ActiveStorage from "@rails/activestorage"
 import "channels"
-import "./posts"
 import "./changeRaidoForReport"
 
 // トースターを閉じるクラス
@@ -243,6 +242,91 @@ function setupPopup() {
   });
 }
 
+/**
+ * CodeMirrorライブラリをカスタマイズする
+ * @returns {void}
+ */
+function setupCodeMirror() {
+  const textareas = document.querySelectorAll("textarea.js-code-textarea");
+  if(textareas.length === 0)return;
+  const INITIALIZED_FLAG = "flag-cm-initialized";
+
+  textareas.forEach(textarea => {
+    if(!textarea.classList.contains(INITIALIZED_FLAG)){
+      CodeMirror.fromTextArea(textarea, {
+        lineNumbers: true,
+        indentUnit: 2,
+        tabSize: 2,
+        indentWithTabs: false,
+        extraKeys: {
+          Tab: cm => cm.execCommand("indentMore"),
+          "Shift-Tab": cm => cm.execCommand("indentLess")
+        }
+      });
+    }
+    textarea.classList.add(INITIALIZED_FLAG);
+  });
+}
+
+/**
+ * CodeMirrorの取得が遅れる場合があったため、ポーリングをかける
+ * @param {number} 試行回数上限
+ */
+function pollingCodeMirror(retryCount = 10) {
+  if (typeof CodeMirror === "undefined") {
+    if (retryCount <= 0) return;
+    setTimeout(() => pollingCodeMirror(retryCount - 1), 100);
+    return;
+  }
+  setupCodeMirror();
+}
+
+
+/**
+ * 動的にフォームのフィールドを追加・削除する
+ * @param {function} afterAddedCallBack - フィールド追加後に実行する関数
+ * @returns {void}
+ */
+function setupAddFields(afterAddedCallBack) {
+  const container = document.getElementById("js-fields-container");
+  const template = document.getElementById("js-field-template");
+  const addBtn = document.getElementById("js-add-field");
+  if (!container || !template || !addBtn) return;
+
+  // name="post[モデル_attributes][数字][属性名]" の「数字」に対応した番号
+  let index = document.querySelectorAll(".js-field-block").length;
+
+  const templateHTML = template.innerHTML;
+
+  // フィールド追加処理
+  addBtn.addEventListener("click", () => {
+    const html = templateHTML.replace(/JS_INDEX/g, index);
+    container.insertAdjacentHTML("beforeend", html);
+    afterAddedCallBack()
+    index++;
+  });
+
+  // フィールド削除処理
+  container.addEventListener("click", (e) => {
+    // イベントの伝播を利用して、js-remove-field要素の擬似押下を行う。
+    // => フィールドを追加するたびに、js-remove-field要素にクリックイベントを仕込むと記述量が多くなってしまうから。
+    const removeBtn = e.target.closest(".js-remove-field");
+    if (removeBtn) {
+      const block = e.target.closest(".js-field-block");
+      const destroyInput = block.querySelector('input[name*="_destroy"]')
+      if(destroyInput) {
+        // accepts_nested_attributes_for :モデル, allow_destroy: trueにより、値が"1"になると削除する合図を送る
+        destroyInput.value = "1"; 
+
+        block.style.display = "none";
+      } else {
+        // 万が一 _destroy が存在しないときはDOMごと削除
+        block.remove()
+      };
+    };
+  });
+}
+
 Rails.start()
 Turbolinks.start()
 ActiveStorage.start()
@@ -259,4 +343,6 @@ document.addEventListener("turbolinks:load", () => {
   setupTab();
   chatScrollBottom()
   setupPopup()
+  setupAddFields(setupCodeMirror)
+  pollingCodeMirror();
 });
