+++ app/views/public/posts/_form_fields.html.erb
@@ -14,16 +14,16 @@
   <section class="u-sec">
     <div class="u-inner-4">
       <h2 class="u-title-1">コード（必須）</h2>
-      <div id="js-code-fields-container">
+      <div id="js-fields-container">
         <%# fields_for: 入れ子のモデルのフォームを生成するためのヘルパー。引数は子モデルをシンボルで取り、ブロック引数は、子モデルに対するフォームビルダーオブジェクトとして扱う。 %>
         <%= f.fields_for :codes do |code_fields| %>
           <%= render 'code_fields', f: code_fields %>
         <% end %>
       </div>
       <div class="u-btn-wrap u-mt-30">
-        <button type="button" id="js-add-code-field" class="u-btn opt-size-s" aria-label="フィールドを追加する"><i class="fa-solid fa-plus"></i> フィールドを追加する</button>
+        <button type="button" id="js-add-field" class="u-btn opt-size-s" aria-label="フィールドを追加する"><i class="fa-solid fa-plus"></i> フィールドを追加する</button>
       </div>
-      <template id="js-code-field-template">
+      <template id="js-field-template">
         <%# raw() => テンプレで使うコードをエスケープすると、文字列のまま出力されてしまう。なのでそれを防止する。セキュリティ面: このケースはユーザーが入力する使い方をしていないので、問題ない。 %>
         <%# gsub("\n", "") => テンプレ内部の改行を取り除く %>
         <%# jsで扱うHTML文字列を1行にしてバグなどを防ぐ %>
