+++ app/views/public/users/edit.html.erb
@@ -139,6 +139,14 @@
                 </div>
               </div>
             <% end %>
+
+            <template>
+              
+            </template>
+
+            <div class="u-btn-wrap u-mt-30">
+              <button type="button" id="js-add-field" class="u-btn opt-size-s" aria-label="フィールドを追加する"><i class="fa-solid fa-plus" aria-hidden="true"></i> フィールドを追加する</button>
+            </div>
           </div>
         </section>
       </div>
